+++ tests/test_path.py
@@ -60,9 +60,10 @@ def test_extend(setup_test_environment):
 
     # EMPTY PATH
     assert Path.extend("") is None
-    assert Path.extend(None) is None  # type: ignore[assignment]
     with pytest.raises(PathNotFoundError, match="Path is empty."):
         Path.extend("", raise_error=True)
+    with pytest.raises(TypeError, match="parameter must be a string"):
+        Path.extend(None, raise_error=True)  # type: ignore[assignment]
 
     # FOUND IN STANDARD LOCATIONS
     assert Path.extend("file_in_cwd.txt") == str(env["cwd"] / "file_in_cwd.txt")